+++ Services/UsuarioService.cs
@@ -1,5 +1,4 @@
-﻿using CareWithLoveApp.Data;
-using CareWithLoveApp.Models.Entities;
+﻿using CareWithLoveApp.Models.Entities;
 using CareWithLoveApp.Repositories;
 using System;
 using System.Collections.Generic;
@@ -17,6 +16,11 @@ namespace CareWithLoveApp.Services
 
         public Usuario? ObterUsuarioPorId(Guid id)
         {
+            if (id == Guid.Empty)
+            {
+                throw new ArgumentException("O ID do usuário não pode ser vazio.", nameof(id));
+            }
+
             return _usuarioRepository.GetById(id);
         }
 
@@ -27,22 +31,41 @@ namespace CareWithLoveApp.Services
 
         public void CriarUsuario(Usuario usuario)
         {
-            // Aqui você pode aplicar regras de negócio
+            if (usuario == null)
+            {
+                throw new ArgumentNullException(nameof(usuario), "O usuário não pode ser nulo.");
+            }
+
             _usuarioRepository.Add(usuario);
         }
 
         public void AtualizarUsuario(Usuario usuario)
         {
+            if (usuario == null)
+            {
+                throw new ArgumentNullException(nameof(usuario), "O usuário não pode ser nulo.");
+            }
+
             _usuarioRepository.Update(usuario);
         }
 
         public void ExcluirUsuario(Guid id)
         {
+            if (id == Guid.Empty)
+            {
+                throw new ArgumentException("O ID do usuário não pode ser vazio.", nameof(id));
+            }
+
             _usuarioRepository.Delete(id);
         }
 
         public void AdicionarUsuario(Usuario usuario)
         {
+            if (usuario == null)
+            {
+                throw new ArgumentNullException(nameof(usuario), "O usuário não pode ser nulo.");
+            }
+
             _usuarioRepository.Add(usuario);
         }
     }
